--- customer/src/main/java/nl/softwarestrijders/waiter/customer/core/application/command/UpdateProduct.java
@@ -1,4 +1,0 @@
-package nl.softwarestrijders.waiter.customer.core.application.command;
-
-public class UpdateProduct {
-}
