+++ customer/src/main/java/nl/softwarestrijders/waiter/customer/core/application/command/consume/UpdatedProduct.java
@@ -1,0 +1,4 @@
+package nl.softwarestrijders.waiter.customer.core.application.command.consume;
+
+public class UpdatedProduct {
+}
