+++ customer/src/main/java/nl/softwarestrijders/waiter/customer/core/application/command/produce/AddOrderToCustomer.java
@@ -1,0 +1,4 @@
+package nl.softwarestrijders.waiter.customer.core.application.command.produce;
+
+public class AddOrderToCustomer {
+}
