--- customer/src/main/java/nl/softwarestrijders/waiter/customer/core/application/query/CreateCustomer.java
@@ -1,4 +1,0 @@
-package nl.softwarestrijders.waiter.customer.core.application.query;
-
-public class CreateCustomer {
-}
